Move FetchState enum to actions
