+++ packages/synapse-interface/slices/portfolio/hooks.ts
@@ -3,8 +3,8 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 import { getAccount, Address } from '@wagmi/core'
 
 import { AppDispatch, RootState } from '@/store/store'
+import { FetchState } from './actions'
 import { useAppSelector } from '@/store/hooks'
-import { FetchState } from './reducer'
 import {
   fetchPortfolioBalances,
   NetworkTokenBalancesAndAllowances,
